Issue #SB-14499 fix: Code review comments changes
