+++ platform-modules/batch-models/src/test/resources/application.conf
@@ -96,6 +96,10 @@ assessment.metrics.cloud.provider="azure"
 assessment.metrics.content.index="compositesearch"
 assessment.metrics.es.alias="cbatch-assessment"
 assessment.metrics.es.index.prefix="cbatch-assessment-"
+course.upload.reports.enabled=false
+course.es.index.enabled=false
+
+
 
 # for only testing uploads to blob store
 azure_storage_key=""
