+++ platform-modules/batch-models/src/test/scala/org/ekstep/analytics/job/TestAssessmentMetricsJob.scala
@@ -3,7 +3,7 @@ package org.ekstep.analytics.job
 import org.apache.spark.sql.{DataFrame, SparkSession}
 import org.ekstep.analytics.job.AssessmentMetricsJob.saveReport
 import org.ekstep.analytics.model.SparkSpec
-import org.ekstep.analytics.util.AssessmentReportUtil
+import org.ekstep.analytics.util.{AssessmentReportUtil, ESUtil}
 import org.scalamock.scalatest.MockFactory
 import org.sunbird.cloud.storage.conf.AppConf
 
@@ -113,6 +113,8 @@ class TestAssessmentMetricsJob extends SparkSpec(null) with MockFactory {
     assert(AppConf.getConfig("assessment.metrics.cloud.provider").isEmpty === false)
     assert(AppConf.getConfig("course.metrics.cloud.container").isEmpty === false)
     assert(AppConf.getConfig("assessment.metrics.temp.dir").isEmpty === false)
+    assert(AppConf.getConfig("course.upload.reports.enabled").isEmpty === false)
+    assert(AppConf.getConfig("course.es.index.enabled").isEmpty === false)
   }
 
   "TestAssessmentMetricsJob" should "Ensure CSV Report Should have all proper columns names and required columns values" in {
@@ -287,7 +289,7 @@ class TestAssessmentMetricsJob extends SparkSpec(null) with MockFactory {
     val contentESIndex = AppConf.getConfig("assessment.metrics.content.index")
     assert(contentESIndex.isEmpty === false)
     val contentList = List("do_112835336280596480151", "do_112835336960000000152")
-    val contentDF = AssessmentReportUtil.getContentNames(spark, contentList)
+    val contentDF = ESUtil.getContentNames(spark, contentList, AppConf.getConfig("assessment.metrics.content.index"))
     assert(contentDF.count() === 2)
   }
